Refactor code to simplify model building and bugfix in the formulation.
Before reactions where inferred from the pkn graph structure.
Now, reactions are read directly from the pkn.reactions attribute.

A bug in the formulation of one of the constraints has been fixed. There
was a "-1" being added outside pulp.lpSum(). Now it is included inside
pulp.lpSum(), correctly matching the mathematical formulation.

The name of the variables has been changed to better match the one in the
mathematical formulation.