Name the constraints.
The constraints in the model are now named to facilitate their
identification. The namce starts with 'Cx_', where 'C' stands for
constraint and x is a number (the number of the equation representing the
constraint in the paper Mitsos et al. 2009). After 'Cx_', the corresponding
i, j or k values are appended, separated by an underscore ('_'). Note that
some values of i, j or k may have underscores in their names. The order in
which this i, j or k values appear is the same in which their are written
in the equations of the paper.